// industry-verticals/retail/src/types/styleFlags.ts
export enum CommonStyles {
  HideAccentLine = 'hide-accent-line',
}

export enum LayoutStyles {
  Reversed = 'reversed',
}

export enum PromoFlags {
  PromoReversed = 'promo-reversed',
  ShowMultipleImages = 'show-multiple-images',
  HidePromoShapes = 'hide-promo-shapes',
  HidePromoShadows = 'hide-promo-shadows',
  HidePromoQuotes = 'hide-promo-quotes',
}

export enum HeroBannerStyles {
  HideGradientOverlay = 'hide-gradient-overlay',
  WithPlaceholder = 'with-placeholder',
  ScreenLayer = 'screen-layer',
  ImageFitCover = 'image-fit-cover',
}